:truck: Timer 구조 변경

+++ G-11/Game/Timer/Timer.cpp
@@ -1,14 +1,15 @@
 ﻿#include "Timer.h"
 
-Timer::Timer(float targetTime)
+Timer::Timer(double targetTime)
     : targetTime(targetTime)
 {
+    
 }
 
 void Timer::Tick(float deltaTime)
 {
     // 경과시간 업데이트.
-    elapsedTime += deltaTime;
+    elapsedTime += static_cast<double>(deltaTime);
 }
 
 void Timer::Reset()
