+++ G-11/Game/Timer/Timer.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 
 /*
-* 작성자: 장세윤 강사님. 2025.07.31
+* 작성자: 장세윤. 2025.07.31
 * 타이머 클래스
 * - 초 단위로 시간 계산이 필요할 때 사용.
 * - 예: 3초가 지났는지 등을 확인 할 때 사용함.
@@ -10,7 +10,7 @@
 class Timer
 {
 public:
-	Timer(float targetTime = 0.0f);
+	Timer(double targetTime = 0.0);
 
 	// 초시계의 시간이 흘러가도록 하기 위해 필요함.
 	void Tick(float deltaTime);
@@ -25,6 +25,6 @@ public:
 	void SetTargetTime(float newTargetTime);
 
 private:
-	float elapsedTime = 0.0f;
-	float targetTime = 0.0f;
+	double elapsedTime = 0.0;
+	double targetTime = 0.0;
 };
